+++ src/components/DateFormat/index.js
@@ -1,6 +1,8 @@
 import React, { useState } from 'react';
-import DatePicker from 'react-datepicker';
+import DatePicker, { registerLocale, setDefaultLocale } from 'react-datepicker';
 import 'react-datepicker/dist/react-datepicker.css';
+import th from 'date-fns/locale/th';
+registerLocale('th', th);
 
 // CSS Modules, react-datepicker-cssmodules.css
 // import 'react-datepicker/dist/react-datepicker-cssmodules.css';
@@ -10,6 +12,7 @@ const CustomDateFormat = () => {
 
   return (
     <DatePicker
+      locale="th"
       dateFormat="yyyy/MM/dd"
       selected={startDate}
       onChange={(date) => setStartDate(date)}
